requirements.txt
opencv-python>=4.11.0.86
opencv-contrib-python>=4.11.0.86
scikit-learn>=1.7.0
numpy==1.26.4
matplotlib>=3.10.3
pillow>=11.3.0
ultralytics>=8.3.163
mediapipe==0.10.21
seaborn>=0.13.2
joblib>=1.5.1